requirements.txt
GitPython~=3.1.40
PyYAML~=6.0.2
accelerate~=1.2.1
bitsandbytes~=0.45.3
build~=1.2.2.post1
click~=8.1.8
configparser~=7.1.0
datasets~=3.2.0
deepspeed~=0.16.4
evaluate~=0.4.3
hydra-core~=1.3.2
ipywidgets~=8.1.5
joblib~=1.4.2
jupyterlab~=4.3.5
llamafactory~=0.9.2
mlflow~=2.20.3
networkx~=3.4.2
nltk~=3.9.1
numba~=0.60.0
numpy~=1.26.3
omegaconf~=2.3.0
openai~=1.65.1
openpyxl~=3.1.5
pandas~=2.1.4
peft~=0.12.0
psutil~=7.0.0
requests~=2.32.3
scikit-learn~=1.6.1
sharedmem~=0.3.8
spacy~=3.8.4
thefuzz~=0.22.1
tqdm~=4.67.1
transformers~=4.49.0
trl~=0.9.6
urllib3~=2.3.0
unsloth~=2025.2.15